requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt requirements/app.in requirements/test.in
#
click==6.6
coverage==4.2             # via pytest-cov
mypy-lang==0.4.6
pathlib==1.0.1
py==1.4.31                # via pytest
pygraphviz==1.3.1
pytest-cov==2.4.0
pytest-mock==1.5.0
pytest==3.0.4
pyyaml==3.12
